violation/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python
